pyproject.toml
[project]
name = "nlp-advanced-2025-epita"
version = "0.1.0"
description = "Overhearing Agents in RP using Local LLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.19.0,<2.0.0",  # TTS requires numpy<2.0
    "openai-whisper>=20250625",
    "faster-whisper>=1.1.0",
    "piper-tts>=1.3.0",
    "TTS>=0.22.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.12.0",
    "scipy>=1.11.0",
    "torch>=2.6.0,<3.0.0",  # sm_120 GPU needs torch 2.6+ (build with Blackwell support)
    "torchcodec>=0.1.0",    # required by XTTS load_with_torchcodec
    "langchain[all]>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.1.0",
    "silero-vad>=6.2.0",
    "pyaudio>=0.2.14",
    # RAG module dependencies
    "chromadb>=0.4.15",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "transformers>=4.30.0,<4.40.0",  # XTTS v2 requires transformers<4.40
    # Testing
    "pytest>=8.0.0",
]
